+++ samples/wavlib/src/main/cpp/io/stream/InputStream.h
@@ -16,9 +16,11 @@
 #ifndef _IO_STREAM_INPUTSTREAM_H_
 #define _IO_STREAM_INPUTSTREAM_H_
 
+#include <cstdint>
+
 namespace wavlib {
 
-/*
+/**
  * An interface declaration for a stream of bytes. Concrete implements for File and Memory Buffers
  */
 class InputStream {
@@ -26,34 +28,34 @@ public:
     InputStream() {}
     virtual ~InputStream() {}
 
-    /*
-     * Retreive the specified number of bytes and advance the read position.
+    /**
+     * Retrieve the specified number of bytes and advance the read position.
      * Returns: The number of bytes actually retrieved. May be less than requested
      * if attempt to read beyond the end of the stream.
      */
-    virtual long read(void *buff, long numBytes) = 0;
+    virtual int32_t read(void *buff, int32_t numBytes) = 0;
 
-    /*
-     * Retreive the specified number of bytes. DOES NOT advance the read position.
+    /**
+     * Retrieve the specified number of bytes. DOES NOT advance the read position.
      * Returns: The number of bytes actually retrieved. May be less than requested
      * if attempt to read beyond the end of the stream.
      */
-    virtual long peek(void *buff, long numBytes) = 0;
+    virtual int32_t peek(void *buff, int32_t numBytes) = 0;
 
-    /*
+    /**
      * Moves the read position forward the (positive) number of bytes specified.
      */
-    virtual void advance(long numBytes) = 0;
+    virtual void advance(int32_t numBytes) = 0;
 
-    /*
+    /**
      * Returns the read position of the stream
      */
-    virtual long getPos() = 0;
+    virtual int32_t getPos() = 0;
 
-    /*
+    /**
      * Sets the read position of the stream to the 0 or positive position.
      */
-    virtual void setPos(long pos) = 0;
+    virtual void setPos(int32_t pos) = 0;
 };
 
 } // namespace wavlib
